This commit inclues the following changes:
Expects (and calculates Fresnel reflection based on) medium index instead of matrix index. Using matrix index here is somewhat incorrect, as we have been using the matrix index to mean the medium index throughout the development of this code. This change doesn't actually affect behaviour, it only changes the name of this variable.

calc_reflectance function renamed to calc_refl_trans and now outputs (reflectance, transmittance) as a tuple instead of just reflectance.

Transmittance and reflectance are now calculated relative to the sum of absorbed + transmitted + reflected light, instead of incident light. This means that light with no determined end state is not included in this normalization and is assumed to be absorbed/transmitted/reflected in the same fraction as light with a determined end state. Light with no determined end state includes trajectories that do not exit the sample or light that is fresnel reflected back into the sample when it tries to exit. These two light outcomes were previously assumed to be not reflected (ie either absorbed or transmitted), but are now assumed to have the same outcome probabilities as the light that is accounted for.

Averaging over trajectories is only performed as the very last step, as opposed to averaging over Fresnel reflection at each interface separately. For example, this means that a trajectory that is incident on the sample at a large angle (high reflection) will have the fresnel reflection correction applied to its outcome only, and not averaged over all trajectories.

Calculation of refraction when leaving the sample is no longer computed. I plan to add this back in, although it seems strange to include this but not account for refraction upon entering the sample. Eventually, this incident refraction should also be added.

Comparison of output angles to detection angle is no longer performed. I plan to add this back in.ework code to calculate reflectance and transmittanceRework code to calculate reflectance and transmittance