Update how calc_refl_trans determines trajectory outcomes. Changes are as follows:
Trajectories that TIR are reflected back into the sample and continued.

Stuck trajectories are assumed to have forgotten their initial position and direction and so are distributed evenly between reflection and transmission.

Portions of trajectories that are fresnel reflected back into the sample upon exit (not TIR) are treated differently than stuck trajectories. They are considered as new trajectories being introduced at the interface where the fresnel reflection took place, and are reflected or transmitted in the same ratios as all initial trajectories (which enter from the upper interface only).

Updated tests to confirm the new code behaves as intended.