routed all differential cross-section calcs through Model methods
changed montecarlo.phase_function() to use Model class methods instead of using
the generic differential_cross_section() method in model.py.  Left in some
assert statements to check that the differential cross sections returned by the
Model objects were the same as those of the generic function.  They are to
within 2e-11 rtol.  The deviation seems to be from rounding error from
differences in the order of converting quantities, for example.  It should now
be possible to remove the differential_cross_section() generic function.