added Model.phase_function() method and tests
Note that phase function is now normalized to the total scattering
cross-section.  Previously was normalized to the sum of the differential
scattering cross-sections, but this is incorrect because it doesn't scale with
the cross-section (sum of elements doesn't include the sin(theta) factor).  The
faulty normalization didn't present a problem previously because the phase
function was normalized again to provide a probability distribution.