added number_density property to Model objects and added tests
This change entailed changing FormStructureModel to optionally take a particle
and/or volume_fraction in its constructor.  These properties are necessary to
calculate absorption and/or transport lengths from the FormStructureModel, which
would otherwise not be aware of the particle index or number density.  Changed
HardSpheres and PolydisperseHardSpheres to pass the volume_fraction argument to
the FormStructureModel constructor.  number_density method must be overriden in
derived classes of FormStructureModel