fixed bug in specular reflection handling in model.reflection()
Previous code assumed that the detector always included the specular angle, and
so fresnel reflection was always included.  New code checks to see if the
detector includes the specular angle as a condition for adding fresnel
reflection from the sample-medium interface.  Updated test_theta_refraction to
test for correct behavior.