changed unit registry to allow subclassing pint.Quantity
Needed to subclass pint.UnitRegistry and then add a new subclassed pint.Quantity
object to it.  The idea is to fix some bugs in pint.Quantity and add methods.
To prevent two unit registries from being created when using pymie and structcol
together, we use pint.set_application_registry() in structcol and
pint.get_application_registry() in pymie.  pymie will then detect that the
structcol unit registry is being used and adopt it, if structcol is imported
first.