removed line in detector.py that takes abs(n_sample); updated tests
The warnings cited as the reason for taking the absolute value no longer appear
to be a problem, and we should avoid taking absolute values of complex
quantities except where necessary. Most of the updated expected values for the
tests are small changes from previous values, and many were from tests based on
notebooks.