QtyTrajectory->NumpyTrajectory; moved conversion to detector funcs
Conversion from Dataset to NumpyTrajectory is now handled by calc_refl_trans()
and other functions in the detector suite, which should make it easier to
convert these functions to handle Dataset.  Because it is no longer necessary to
convert manually, the test functions no longer need to have an explicit
conversion to NumpyTrajectory.