modified calc_refl_trans() to take Simulation object
This pares down the parameter list for this function and for
run_sphere_fresnel_traj() because the sim object contains all information from
the simulation, including the model used, the indices of refraction, and the
phase function.  Had to change the values for a couple of tests involving
surface roughness.  Previously these tests were using n_tir=n_sample in
calc_refl_trans(), since n_matrix and n_particle were not specified.  Now that
the sim object is given to calc_refl_trans, n_matrix and n_particle are
available to be calculated from the model object, and as a result, n_tir can be
calculated using the roughness correction.