created ConstantIndex object and removed Index.constant class method
Previously one had to write sc.Index.constant to get a constant index, which was
a little confusing because we generally write sc.index (with a lower case "i")
to access pre-defined Index objects in the refractive_index module.
sc.ConstantIndex now replaces sc.Index.constant, which is more logical and
consistent with how EffectiveIndex (the other subclass of Index) is defined.