fixed bug in rotate_refract() and updated tests
The rotation formulas used in this function assume that the rotation axis vector
<u, v, w> has length 1, but some of the calling functions and tests do not
explicitly enforce this constraint. So I changed the function to normalize the
axis vector before applying the rotation. This change leads to changes in some
of the test results for test_detector_sphere(). Although these changes are
small, they are outside the error tolerances (the changes have rtol 10^-2 to
10^-5), so I updated the expected values in the affected tests