change rotate_refract() to use scipy.spatial.transform.Rotation
This change makes it easier to specify and calculate the rotation matrix. Also
the previous version of rotate_refract() used a rotation formula for rotation
about an arbitrary axis that can be offset from the origin. However, the
function was used only for rotation axes that go through the origin, and indeed
it translated the offset to the origin anyway. So I simplified the function by
removing the offset and using a standard rotation matrix.