changed select_events() to enable vectorized Monte Carlo simulations
This change means that we no longer drop trajectories that are not selected. Had
to change several functions where it was assumed that only selected trajectories
were present in a DataArray. But this change also simplifies the code; we no
longer have to use reindex_like() or combine_first() before adding or
multiplying DataArrays, since all DataArrays now have all trajectories. Added
test to confirm that vectorization of MC simulations over wavelength works and
gives values within 5% of looped calculations.